driver: use ALARM log for Tag errors.
This change makes `state` and `relay` post a HOST-144 (Comm Tag error)
error to the central ALARM log of the controller upon detecting a problem
with either the configuration or state of a server Tag during
initialisation.

The error is posted with ABORT priority, causing the controller to abort
the program(s). Additional information is printed on the User Menu.