driver: add explicit socket close on ABORT. Fix #27.
This change lets the state and trajectory relay components installs
an ABORT condition handler that will trap ABORT signals, and signal
to the respective applications that the user has requested the
programs to stop.

The handler will allow both programs to properly close the file
descriptor opened on the TCP socket, thereby properly shutting down
the socket connection itself.