driver: monitor ros_relay program status. Fix #108.
Uses GET_TSK_INFO(..) routine to determine whether both the trajectory
relay (ros_relay) and the motion statemachine (ros_movesm) are actually
running. If they are not, motion is not possible and this is reflected in
the value of the 'motion possible' field of the 'robot status' message.

Also: both the ros_state and ros_relay KAREL program are now started from
the main TP program using RUN(..). This makes it easier to resume
ros / ros_movesm if it ever gets into PAUSED state.